--- data/cache/dao/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
-
-}
-
-android {
-    namespace 'com.goldenspear.android.aihealth.band.medical.dao'
-
-    defaultConfig {
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-}
-
-dependencies {
-    implementation project(path: ':domain:model')
-    kapt libs.room.compiler
-
-    api libs.kotlin.coroutines
-
-    api libs.bundles.paging
-
-}
