--- data/cache/db/.gitignore
@@ -1,2 +1,0 @@
-/build
-/schemas/com.jfmr.ac.test.data.cache.db.RickAndMortyDB/1.json
