--- data/cache/db/build.gradle
@@ -1,34 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
-
-}
-
-android {
-    namespace 'com.jfmr.ac.test.data.cache.db'
-
-    defaultConfig {
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-        javaCompileOptions {
-            annotationProcessorOptions {
-                arguments += ["room.schemaLocation": "$projectDir/schemas".toString()]
-            }
-        }
-    }
-}
-
-dependencies {
-    implementation project(path: ':domain:model')
-    implementation(project(path: ":data:open"))
-    implementation(project(path: ":data:cache:dao"))
-
-    implementation libs.bundles.hilt
-    kapt libs.bundles.hiltkapt
-
-    kapt libs.room.compiler
-
-    api libs.paging.room
-    implementation libs.google.gson
-
-}
