+++ data/remote/build.gradle
@@ -13,8 +13,7 @@ kapt {
 dependencies {
     implementation(project(path: ":domain:model"))
     implementation(project(path: ":data:open"))
-    implementation(project(path: ":data:cache:dao"))
-    implementation(project(path: ":data:cache:db"))
+    implementation(project(path: ":data:cache"))
 
     implementation libs.bundles.hilt
     kapt libs.bundles.hiltkapt
