+++ data/repository/build.gradle
@@ -12,7 +12,7 @@ kapt {
 }
 dependencies {
     implementation(project("path": ":domain:model"))
-    implementation(project("path": ":domain:repository:open"))
+    implementation(project("path": ":domain:repository"))
     implementation(project(path: ":data:remote"))
     implementation(project(path: ":data:open"))
