+++ domain/repository/build.gradle
@@ -1,0 +1,14 @@
+plugins {
+    id 'com.android.library'
+    id 'kotlin-android'
+
+}
+
+android {
+    namespace = "com.jfmr.domain.ac.test.repository"
+}
+
+dependencies {
+    api(project(path: ":domain:model"))
+    implementation(project(path: ":data:cache"))
+}
