--- domain/repository/open/build.gradle
@@ -1,14 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-
-}
-
-android {
-    namespace = "com.jfmr.domain.ac.test.repository.open"
-}
-
-dependencies {
-    api(project(path: ":domain:model"))
-    implementation(project(path: ":data:cache:db"))
-}
