--- domain/repository/open/src/main/java/com/jfmr/ac/test/domain/repository/open/character/CharacterRepository.kt
@@ -1,15 +1,0 @@
-package com.jfmr.ac.test.domain.repository.open.character
-
-import androidx.paging.PagingData
-import com.jfmr.ac.test.domain.model.character.DomainCharacter
-import com.jfmr.ac.test.domain.model.character.DomainResult
-import kotlinx.coroutines.flow.Flow
-
-interface CharacterRepository {
-
-    fun characters(): Flow<PagingData<DomainCharacter>>
-
-    suspend fun getCharacterById(id: Int): DomainResult<DomainCharacter>
-
-    suspend fun updateCharacter(character: DomainCharacter): DomainCharacter
-}
