+++ domain/usecase/di/build.gradle
@@ -14,7 +14,7 @@ dependencies {
     api(project(":data:open"))
     api(project(":data:di"))
     api(project(":domain:model"))
-    api(project(":domain::usecase:implementation"))
+    implementation(project(":domain:usecase:implementation"))
     api(project(":domain:usecase:open"))
 
     implementation libs.bundles.hilt
