+++ domain/usecase/implementation/src/main/java/com/jfmr/domain/usecase/implementation/character/CharacterDetailInteractor.kt
@@ -3,7 +3,7 @@ package com.jfmr.domain.usecase.implementation.character
 import com.jfmr.ac.test.data.di.QCharacterRepository
 import com.jfmr.ac.test.domain.model.character.DomainCharacter
 import com.jfmr.ac.test.domain.model.error.DomainError
-import com.jfmr.ac.test.domain.repository.open.character.CharacterRepository
+import com.jfmr.ac.test.domain.repository.character.CharacterRepository
 import com.jfmr.ac.test.domain.usecase.open.character.CharacterDetailUseCase
 import javax.inject.Inject
