+++ domain/usecase/implementation/src/main/java/com/jfmr/domain/usecase/implementation/character/CharactersInteractor.kt
@@ -1,7 +1,7 @@
 package com.jfmr.domain.usecase.implementation.character
 
 import com.jfmr.ac.test.data.di.QCharacterRepository
-import com.jfmr.ac.test.domain.repository.open.character.CharacterRepository
+import com.jfmr.ac.test.domain.repository.character.CharacterRepository
 import com.jfmr.ac.test.domain.usecase.open.character.CharactersUseCase
 import javax.inject.Inject
