+++ domain/usecase/implementation/src/main/java/com/jfmr/domain/usecase/implementation/character/UpdateCharacterInteractor.kt
@@ -2,7 +2,7 @@ package com.jfmr.domain.usecase.implementation.character
 
 import com.jfmr.ac.test.data.di.QCharacterRepository
 import com.jfmr.ac.test.domain.model.character.DomainCharacter
-import com.jfmr.ac.test.domain.repository.open.character.CharacterRepository
+import com.jfmr.ac.test.domain.repository.character.CharacterRepository
 import com.jfmr.ac.test.domain.usecase.open.character.UpdateCharacterUseCase
 import javax.inject.Inject
