+++ presentation/ui/src/main/java/com/jfmr/ac/test/presentation/ui/character/list/viewmodel/CharacterListViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.jfmr.ac.test.domain.model.character.DomainCharacter
-import com.jfmr.ac.test.domain.usecase.di.RetrieveItemsQualifier
-import com.jfmr.ac.test.domain.usecase.di.UpdateCharacterQualifier
+import com.jfmr.ac.test.domain.usecase.di.GetCharacters
+import com.jfmr.ac.test.domain.usecase.di.UpdateCharacter
 import com.jfmr.ac.test.domain.usecase.open.character.CharactersUseCase
 import com.jfmr.ac.test.domain.usecase.open.character.UpdateCharacterUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -16,8 +16,8 @@ import javax.inject.Inject
 
 @HiltViewModel
 class CharacterListViewModel @Inject constructor(
-    @RetrieveItemsQualifier private val charactersUseCase: CharactersUseCase,
-    @UpdateCharacterQualifier private val updateCharacterUseCase: UpdateCharacterUseCase,
+    @GetCharacters private val charactersUseCase: CharactersUseCase,
+    @UpdateCharacter private val updateCharacterUseCase: UpdateCharacterUseCase,
 ) : ViewModel() {
 
     var pager: Flow<PagingData<DomainCharacter>> = charactersUseCase.characters().cachedIn(viewModelScope)
