+++ settings.gradle
@@ -93,7 +93,7 @@ include ':domain:usecase:open'
 include ':domain:usecase:implementation'
 include ':domain:usecase:di'
 include ':domain:model'
-include ':domain:repository:open'
+include ':domain:repository'
 
 include ':framework'
 
@@ -102,5 +102,3 @@ include ':data:di'
 include ':data:remote'
 include ':data:cache'
 include ':data:repository'
-include ':data:cache:dao'
-include ':data:cache:db'
